+++ devito/types/dense.py
@@ -18,10 +18,10 @@ from devito.mpi import MPI
 from devito.parameters import configuration
 from devito.symbolics import Add, FieldFromPointer
 from devito.finite_differences import Differentiable, generate_fd_shortcuts
-from devito.tools import (EnrichedTuple, ReducerMap, ArgProvider, as_tuple,
-                          flatten, is_integer, ctypes_to_cstr, memoized_meth,
-                          dtype_to_ctype)
+from devito.tools import (EnrichedTuple, ReducerMap, as_tuple, flatten, is_integer,
+                          ctypes_to_cstr, memoized_meth, dtype_to_ctype)
 from devito.types.dimension import Dimension
+from devito.types.args import ArgProvider
 from devito.types.basic import AbstractCachedFunction
 from devito.types.utils import Buffer, NODE, CELL
